+++ src/components/home/NavSectionItemComponent.vue
