+++ src/components/home/SectionTitleComponent.vue
